requirements.txt
# Core Dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
langchain>=0.0.350
langchain-community>=0.0.20
ctransformers>=0.2.27
groq>=0.4.0

# NLP Libraries
nltk>=3.8.1
spacy>=3.7.0
textstat>=0.7.3

# Advanced NLP and ML
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
pandas>=2.0.0

# Evaluation Metrics
nltk>=3.8.1

# Document Processing
python-docx>=1.1.0
Pillow>=10.1.0

# HTTP Requests
requests>=2.31.0

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Utilities
tqdm>=4.66.1




